split program into several files
