--- gui_trial.py
@@ -1,55 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Apr 10 20:48:35 2020
-
-@author: hviid
-"""
-
-import tkinter as tk
-
-window.destroy()
-
-window = tk.Tk()
-label = tk.Label(text="Name")
-entry = tk.Entry()
-
-label.pack()
-entry.pack()
-name = entry.get()
-
-window.mainloop()
-
-
-name = "asdasd"
-print(name)
-
-
-
-# label = tk.Label(
-#     text="Hello, Tkinter",
-#     fg="white",
-#     bg="black",
-#     width=30,
-#     height=20
-# )
-
-# button = tk.Button(
-#     text="Click me!",
-#     width=25,
-#     height=5,
-#     bg="blue",
-#     fg="yellow",
-# )
-
-# entry = tk.Entry(fg="yellow", bg="blue", width=50)
-
-# label.pack()
-# button.pack()
-# entry.pack()
-# window.mainloop()
-
-
-#%% 
-
-
-
